backend/agent_person/requirements.txt
# AI Agent Person 依赖

# Web 框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# HTTP 客户端
httpx>=0.25.0

# 数据验证
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 日志
loguru>=0.7.0

# 环境变量
python-dotenv>=1.0.0

# 音视频处理（可选）
# pydub>=0.25.1
# opencv-python>=4.8.0

# 阿里云 SDK（可选，如果使用 SDK 而非 REST API）
# alibabacloud-nls-python-sdk
# dashscope
